refactor: 基础组件库目录优化
