+++ packages/elements/src/components/mobile/SimpleImage/SimpleImage.json
