+++ packages/elements/src/panels/container/Blank/BlankPanel.vue
