+++ packages/elements/src/panels/mobile/NavBar/NavBarPanel.vue
