+++ packages/elements/src/panels/mobile/Offer/OfferPanel.vue
