+++ packages/elements/src/panels/mobile/SimpleImage/SimpleImagePanel.vue
