+++ packages/elements/src/panels/public/AtomicAttribute/AtomicAttributeBox.vue
