+++ packages/elements/src/panels/web/SimpleTable/SimpleTablePanel.vue
